ShoppingBackend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==75.6.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
uritemplate==4.1.1
